// src/app/model/installmentDetails.model.ts
export interface installmentDetails {
    id: number
    customerNumber: string
    location: string
    billNumber: string
    tariff: string
    customerName: string
    billAmount: string
    installNumber: number
    installmentPercn: number
    lpsAmount: number
    principleAmount: number
    vatAmount: number
    totalAmount: number
    dueDate: string
  }